+++ verik-importer/src/main/kotlin/io/verik/importer/ast/sv/element/common/SvElement.kt
@@ -24,6 +24,8 @@ abstract class SvElement {
 
     abstract val location: SourceLocation
 
+    var parent: SvElement? = null
+
     inline fun <reified E : SvElement> cast(): E {
         return when (this) {
             is E -> this