+++ verik-importer/src/main/kotlin/io/verik/importer/ast/sv/element/declaration/SvClass.kt
@@ -29,6 +29,10 @@ class SvClass(
 
     override var type = Core.C_Unit.toType()
 
+    init {
+        declarations.forEach { it.parent = this }
+    }
+
     override fun accept(visitor: SvVisitor) {
         visitor.visitClass(this)
     }