+++ verik-importer/src/main/kotlin/io/verik/importer/ast/sv/element/declaration/SvPort.kt
@@ -32,6 +32,10 @@ class SvPort(
 
     override var signature: String? = null
 
+    init {
+        descriptor.parent = this
+    }
+
     override fun accept(visitor: SvVisitor) {
         visitor.visitPort(this)
     }