+++ verik-importer/src/main/kotlin/io/verik/importer/ast/sv/element/declaration/SvProperty.kt
@@ -30,6 +30,10 @@ class SvProperty(
     val isMutable: Boolean,
 ) : SvDeclaration() {
 
+    init {
+        descriptor.parent = this
+    }
+
     override fun accept(visitor: SvVisitor) {
         visitor.visitProperty(this)
     }