+++ verik-importer/src/main/kotlin/io/verik/importer/ast/sv/element/declaration/SvTask.kt
@@ -29,6 +29,10 @@ class SvTask(
 
     override var type = Core.C_Unit.toType()
 
+    init {
+        valueParameters.forEach { it.parent = this }
+    }
+
     override fun accept(visitor: SvVisitor) {
         visitor.visitTask(this)
     }