+++ verik-importer/src/main/kotlin/io/verik/importer/ast/sv/element/declaration/SvTypeAlias.kt
@@ -29,6 +29,10 @@ class SvTypeAlias(
     val descriptor: SvDescriptor
 ) : SvTypeDeclaration() {
 
+    init {
+        descriptor.parent = this
+    }
+
     override fun accept(visitor: SvVisitor) {
         visitor.visitTypeAlias(this)
     }